requirements.txt
cryptography
python-dotenv
pycryptodome